Move SmallTrieVisitcontext to VisitContext
